+++ CHANGELOG.rst
@@ -1,10 +1,10 @@
 Changelog
 =========
 
-0.x
+dev
 ---
-
-- Improved the g-and-k and Gaussian example models' documentation: aligned naming convention, expanded in the summary statistics's descriptions
+- Added new example: the stochastic Lotka-Volterra model
+- Fix methods.bo.utils.minimize to be strictly within bounds
 
 0.6.3 (2017-09-28)
 ------------------